Move packages into pkg/, handler to main/
