+++ pkg/blobutil/url_test.go
@@ -3,7 +3,7 @@ package blobutil_test
 import (
 	"testing"
 
-	"github.com/Azure/custom-script-extension-linux/blobutil"
+	"github.com/Azure/custom-script-extension-linux/pkg/blobutil"
 	"github.com/stretchr/testify/require"
 )
