+++ pkg/preprocess/testdata/dos_with_bom.sh
