+++ pkg/preprocess/testdata/dos_without_bom.sh
