+++ pkg/preprocess/testdata/mslogo.png
