+++ pkg/preprocess/testdata/script_noshebang.py
