+++ pkg/preprocess/testdata/script_shebang
