+++ pkg/preprocess/testdata/utf16_big_endian_with_bom.sh
