+++ pkg/preprocess/testdata/utf16_little_endian_with_bom.sh
