+++ pkg/preprocess/testdata/utf8_with_bom.sh
