+++ pkg/preprocess/testdata/utf8_without_bom.sh
