+++ pkg/seqnum/seqnum_test.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/Azure/custom-script-extension-linux/seqnum"
+	"github.com/Azure/custom-script-extension-linux/pkg/seqnum"
 	"github.com/stretchr/testify/require"
 )
